bootstrap: widen OutcomeId to u16 + retry HTTP 408 in fetcher (#160)
* bootstrap: widen OutcomeId to u16 + retry HTTP 408 in fetcher (closes #159)

The May-13 backfill run aborted with 5 wallet failures: 3 Polymarket
multi-outcome markets emitting `outcomeIndex: 999` overflowing every
u8-typed waypoint, and 2 HTTP 408 Request Timeout responses being
routed to SourceError::Fatal instead of being retried.

Production-side changes are width-localised behind two new conversion
impls on OutcomeId (`TryFrom<i64>` and `From<OutcomeId> for i64`) so
any future widening collapses to a single edit in `ids.rs`. The
`MarketResolution.winning_outcome_id` field is promoted from raw `u8`
to `OutcomeId` to kill the `.0` deref footgun at simulation.rs:421.
The fetcher's status-code dispatch is refactored behind a named
`is_retryable_status` helper so adding future codes (425, etc.) is
one word.

No schema migration; SQLite INTEGER already stores any i64. The
`TradeFill.outcome_id` NDJSON-serialisation surface widens to `u16`
(not OutcomeId) — same judgment applied to `resolution_record`.

Tests cover the bug surfaces end-to-end: OutcomeId(999) JSON
round-trip, cache insert/load with winner=999 and u16::MAX,
is_retryable_status policy, scenario_retry_on_408 +
scenario_exhaust_retries_408 against axum mock, and an
end-to-end scenario_outcome_id_widening covering cache idempotency
and resolution-match equality for the multi-outcome path.

* Address code-review doc drift in retry-policy and scenario header

+++ crates/backtest/tests/scenario_max_signal_price.rs
@@ -81,7 +81,7 @@ fn mkt(idx: u32) -> MarketId {
 fn raw_trade(
     w: WalletAddress,
     market_idx: u32,
-    outcome: u8,
+    outcome: u16,
     day_offset: u32,
     hour_offset: i64,
     side: Side,
